♻️ refactor: 타입 통합
