+++ src/components/Shared/Card/Card.module.scss
@@ -1,0 +1,47 @@
+.card {
+  display: block;
+  border-radius: 10px;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+  color: unset;
+
+  .image-box {
+    border-radius: 10px 10px 0 0;
+    height: 200px;
+    overflow: hidden;
+
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      text-align: center;
+      transition: 0.3s all;
+
+      &:hover {
+        -webkit-transform: scale(1.3);
+        transform: scale(1.3);
+      }
+    }
+  }
+
+  .time-stamp {
+    @include font-style(1.3rem, #666);
+  }
+
+  p {
+    overflow: hidden;
+    color: #000;
+    text-overflow: ellipsis;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    line-height: 1.5em;
+  }
+
+  .card-content {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    width: 100%;
+    padding: 15px;
+  }
+}
