+++ src/pages/SharedPage/index.tsx
@@ -3,7 +3,7 @@ import Header from '../../components/Common/Header';
 import Titlebar from '../../components/Shared/Titlebar';
 import Searchbar from '../../components/Common/Searchbar';
 import apiRequest from '../../utils/apiRequest';
-import CardList from '../../components/Folder/CardList';
+import CardList from '../../components/Shared/CardList';
 import Footer from '../../components/Common/Footer';
 import Loading from '../../components/Common/Loading';
